+++ src/plone/app/fhirfield/indexes/PluginIndexes/FHIRIndex.py
@@ -5,12 +5,41 @@
 # @Version : $Id$
 # All imports here
 from App.special_dtml import DTMLFile
+from plone.app.fhirfield.compat import json
 from plone.app.fhirfield.helpers import validate_index_name
 from plone.app.fhirfield.interfaces import IFhirResourceValue
 from Products.PluginIndexes.FieldIndex.FieldIndex import FieldIndex
 
+import six
+import warnings
+
 
 __author__ = 'Md Nazrul Islam <email2nazrul@gmail.com>'
+
+
+def make_fhir_index_datum(mapping, fhir_json):
+    """ """
+    container = dict()
+    for key, meta in six.iteritems(mapping):
+        if key not in fhir_json:
+            continue
+
+        if 'properties' in meta and fhir_json.get(key):
+            # nested value
+            container[key] = \
+                make_fhir_index_datum(
+                    meta.get('properties'),
+                    fhir_json.get(key))
+
+        else:
+            if meta['type'] == 'string':
+                container[key] = fhir_json.get(key) and \
+                        six.text_type(fhir_json.get(key)) or None
+
+            elif meta['type'] == 'datetime':
+                # XXX: make DateTime object again?
+                container[key] = fhir_json.get(key)
+    return container
 
 
 class FhirFieldIndex(FieldIndex):
@@ -19,6 +48,24 @@ class FhirFieldIndex(FieldIndex):
     query_options = ('query', 'not')
     manage = manage_main = DTMLFile('dtml/manageFhirFieldIndex', globals())
     manage_main._setName('manage_main')
+    mapping = {
+        'id': {
+            'type': 'string',
+        },
+        'meta': {
+            'properties': {
+                'lastUpdated': {
+                    'type': 'datetime',
+                },
+                'versionId': {
+                    'type': 'string',
+                },
+            },
+        },
+        'resourceType': {
+            'type': 'string',
+        },
+    }
 
     def __init__(self, id, ignore_ex=None,
                  call_methods=None, extra=None,
@@ -36,12 +83,36 @@ class FhirFieldIndex(FieldIndex):
         id_ should contains fhir resource name as prefix"""
         validate_index_name(id_)
 
-    def _get_object_datum(self, obj, attr):
-        """ """
+    def _get_object_datum(self, obj, attr, es_datum=False):
+        """Extra param `es_datum` has been added to working
+        Elasticsearch mapping"""
         datum = super(FhirFieldIndex, self)._get_object_datum(obj, attr)
 
+        if es_datum:
+            # No Filter
+            return datum
+
+        if not datum:
+            # Nothing to do
+            return datum
+
+        fhir_value = None
         if IFhirResourceValue.providedBy(datum):
-            datum = datum.as_json()
+            fhir_value = datum.as_json()
+
+        if isinstance(datum, six.string_types):
+            try:
+                fhir_value = json.loads(datum)
+            except ValueError:
+                warnings.warn(
+                    '{0}\'s value must be valid '
+                    'FHIR Resource json. But got-> {1}'.
+                    format(attr, fhir_value),
+                    UserWarning)
+                return
+        if fhir_value:
+            datum = make_fhir_index_datum(self.mapping, fhir_value)
+
         return datum
 
     def unindex_object(self, documentId):